PowerDisplay: Fold RecordingVcpReader into its only consumer
DdcFakes.cs was a file, a static container type, a `using static` import and a
doc comment explaining the container -- all wrapping one fake with one consumer.
That consumer, VcpFeatureProbeServiceTests, already declares its other two
IVcpFeatureReader doubles as private nested classes, so the PR shipped two
conventions for the same thing and the more elaborate one earned nothing.

RecordingVcpReader joins CoordinatedReader and ThrowingReader at the bottom of
the test class. No test method, attribute or assertion changes.
